Fixed problem in new m counting method.

+++ DESDA.h
@@ -48,8 +48,8 @@ class DESDA
     double w_E = 0.99;
     double avg = 0;
 
-    double _maxM = 1000;
-    double _minM = 100;
+    int _maxM = 1000;
+    int _minM = 100;
     int _mE = 0; // Cardinality of observed new objects
 
     std::vector<double> aemEVals = {};
